DsService/pyproject.toml
[project]
name = "dataservice"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "flask>=3.1.2",
    "kafka-python>=2.2.15",
    "langchain>=1.0.1",
    "langchain-community>=0.4",
    "langchain-core>=1.0.0",
    "langchain-google-genai>=3.0.0",
    "pipreqs>=0.5.0",
    "setuptools>=80.9.0",
]
